Add comment about dependents on the internal structure of Calculus.jl
Other packages starts to depend on the internal structure of this
package. I think it is reasonable to add a comment to tell future
developers that might want to change the internal structure which public
packages they will break, so that they can ping the authors.
See: https://github.com/scidom/DualNumbers.jl/pull/1