Fix derivative of airyprime and add more derivatives (#93)
* Fix derivative of airyprime

Contrary to what the docstring says, airy(2, x) is *not* the derivative
of airy(1, x) (it is just the `airybi' function).  The correct
derivative of `airyprime' can be calculated using the Airy equation (and
this is the same done for `airyaiprime').

* Add new derivatives

The derivatives of the inverse error functions can be found at
* http://mathworld.wolfram.com/InverseErf.html
* http://mathworld.wolfram.com/InverseErfc.html

* Fix derivative of invdigamma

My fault, the argument of the `trigamma` function should be the value of
`invdigamma` itself, see
https://en.wikipedia.org/wiki/Inverse_functions_and_differentiation.

* Remove frexp function

This list is only for univariate functions. `frexp` isn't univariate so
it doesn't belong there.